Refactor Levels upload & edit page (#155)
* Change: rename dashboard to manage

* Change: left padding in header navi

* Change: public folder structure

* Change: margin size in Dashboard sidemenu

* Change: bgColor

* Change: bgColor

* Add: upload button in level list page

* Add: form locales

* Add: textarea component

* Add: textarea
